refactor Add server tests and move db tests
