+++ specs/server.test.js
@@ -1,162 +1,115 @@
 const request = require('supertest');
-const createTables = require('../src/db/tablesConfig.js')
 const server = require('../src/server/index.js');  
-
-describe('database tests', () => {
-  let db = require('../src/db/index.js').db; 
-
-  describe('database should include correct tables with correct lengths', () => {
-    //test that the tables are of correct functionality
-    test('should contain parks table', () => {
-      console.log('is this running')
-      return db.query('SELECT * from parks')
-      .then((result) => {
-        console.log('is this runn')
-        expect(result.length).toBe(59);
-     })
-      .catch((err) => {
-        console.log(err)
-        throw err
+describe('server tests', () => {
+  
+  describe('http request tests', () => {
+    // jest expect assertion library https://facebook.github.io/jest/docs/expect.html#content
+    describe('api/parks test', () => {
+      it('responds to /', (done) => {
+      request(server)
+      .get('/api/parks')
+      .end((err, res) => {
+        if (err) {
+          console.log(err);
+          throw err
+        }
+        //need to switch route to res.json to get the test to pass for this test
+        expect(res.status).toEqual(500)
+        expect(res.body).toEqual({})
       })
-    });
-    test('should contain activities table', () => {
-      return db.query('SELECT * from activities')
-      .then((result) => {
-        expect(result.length).toBe(20);
-     })
-      .catch((err) => {
-        console.log(err)
-        throw err
-      })
-    });
-    test('should contain campgrounds table', () => {
-      return db.query('SELECT * from campgrounds')
-      .then((result) => {
-        expect(result.length).toBe(176);
-     })
-      .catch((err) => {
-        console.log(err)
-        throw err
-      })
-    });
-    test('should contain trails table', () => {
-      return db.query('SELECT * from trails')
-      .then((result) => {
-        expect(result.length).toBe(647);
-     })
-      .catch((err) => {
-        console.log(err)
-        throw err
-      })
-    });
-    test('should contain activities_parks table', () => {
-      return db.query('SELECT * from activities_parks')
-      .then((result) => {
-        expect(result.length).toBe(472);
-     })
-      .catch((err) => {
-        console.log(err)
-        throw err
-      })
-    });
-       test('should contain session table', () => {
-      return db.query('SELECT * from session')
-      .then(result => 
-        expect(result.length).toBeGreaterThanOrEqual(0)
-        )
-      .catch((err) => {
-        console.log(err)
-        throw err
-      })
-    });
-  })
-
-  describe('parks table tests', () => {
-
-    test('should contain the correct latitude and longitude', () => {
-      return db.query('SELECT * from parks where parkcode = \'acad\'')
-      .then(result => {
-        const longLat = {latitude: 44.30777545, longitude: -68.30063316};
-        expect(result[0]).toEqual(expect.objectContaining(longLat));
-      })
-      .catch(err => {
-        console.log(err);
-        throw err
+        done();
       })
     })
-  })
 
-  describe('activities table tests', () => {
-    
-    test('should contain the correct activity name', () => {
-      return db.query('SELECT activity FROM activities where id = 10')
-      .then(result => expect(result[0].activity).toEqual('BOATING'))
-      .catch(err => {
-        console.log(err)
-        throw err
+    describe('/api/campgrounds tests', () => {
+      
+      test('responds with campgrounds', (done) => {
+        request(server)
+        .get('/api/campgrounds')
+        .query({parkId: 1})
+        .end((err, res) => {
+          if (err) {
+            console.log(err);
+            throw err
+          }
+          expect(res.status).toEqual(203)
+          expect(res.body.length).toEqual(2)
+        })
+        done();
       })
     })
-  })
 
-  describe('activities_parks table tests', () => {
-    
-    test('should contain all activities from park', () => {
-      return db.query('SELECT * FROM activities_parks where park_id = 1')
-      .then(result => expect(result.length).toEqual(12))
-      .catch(err => {
-        console.log(err)
-        throw err
+    describe('/api/trails', () => {
+
+      test('responds with trails', (done) => {
+        request(server)
+        .get('/api/trails')
+        .query({parkId: 1})
+        .end((err, res) => {
+          if (err) {
+            console.log(err);
+            throw err
+          }
+          expect(res.status).toEqual(200)
+          expect(res.headers['content-type']).toEqual("application/json; charset=utf-8")
+          expect(res.body.length).toEqual(10);
+        })
+        done();
       })
     })
-  })
 
-  describe('trails table tests', () => {
-    
-    test('should contain the correct activity name', () => {
-      return db.query('SELECT * FROM trails where park_id = 1')
-      .then(result => {
-        expect(result.length).toEqual(10);
-        expect(result[0]).toBeInstanceOf(Object);
-      })
-      .catch(err => {
-        console.log(err)
-        throw err
-      })
-    })
-  })
-
-  describe('trails session tests', () => {
-    
-    test('should contain the correct activity name', () => {
-      return db.query('SELECT * FROM session')
-      .then(result => {
-        expect(result.length).toBeDefined();
-
-      })
-      .catch(err => {
-        console.log(err)
-        throw err
+    describe('/api/park tests', () => {
+      test('responds with individual park data', (done) => {
+        request(server)
+        .get('/api/park/')
+        .query({parkcode: 'acad'})
+        .end((err, res) => {
+          if (err) {
+            console.log(err);
+            throw err
+          }
+          expect(res.status).toEqual(200)
+          expect(res.headers['content-length']).toEqual('880')
+          expect(res.body[1][0].name).toEqual('Acadia National Park')
+          expect(res.body[1][0].name).toBeDefined()
+          expect(res.redirect).toEqual(false);
+        })
+        done()
       })
     })
-  })  
 
-})
-
-describe('http request tests', () => {
-  // jest expect assertion library https://facebook.github.io/jest/docs/expect.html#content
-  describe('basic server test', () => {
-    it('responds to /', (done) => {
-    request(server)
-    .get('/')
-    .end((err, res) => {
-      if (err) {
-        console.log(res);
-        throw err
-      }
-      //need to switch route to res.json to get the test to pass for this test
-      expect(res.status).toEqual(200)
-      expect(res.body).toEqual('hello world!!!')})
-      done();
+    describe('/api/shoppingcart', () => {
+      test('responds with session data', (done) => {
+        request(server)
+        .get('/api/shoppingcart')
+        .end((err, res) => {
+          if (err) {
+            console.log(err) 
+            throw err
+          }
+          expect(res.status).toEqual(200)
+          expect(res.redirect).toEqual(false)
+          expect(res.body).toBeInstanceOf(Object)
+        })
+        done()
+      })
     })
+
+    describe('* tests', () => {
+      test('responds with individual park data', (done) => {
+        request(server)
+        .get('/wildcard')
+        .end((err, res) => {
+          if (err) {
+            console.log(err);
+            throw err
+          }
+          expect(res.redirect).toEqual(true)
+        })
+        done()
+      })
+    })
+
   })
+  
 })
